change document arch
