+++ paddle/phi/kernels/funcs/index_elementwise.cu.h
@@ -21,18 +21,12 @@ limitations under the License. */
 
 #include "paddle/phi/backends/gpu/gpu_launch_config.h"
 #include "paddle/phi/backends/gpu/gpu_primitives.h"
-#include "paddle/phi/common/memory_utils.h"
-#include "paddle/phi/common/place.h"
-#include "paddle/phi/core/dense_tensor.h"
 #include "paddle/phi/kernels/funcs/aligned_vector.h"
-#include "paddle/phi/kernels/funcs/math_function.h"
-#include "paddle/phi/kernels/funcs/stride_utils.h"
+#include "paddle/phi/kernels/funcs/index_elementwise_utils.h"
 #include "paddle/phi/kernels/primitive/kernel_primitives.h"
 
 namespace phi {
 namespace funcs {
-
-constexpr int MAX_DIMS = 9;
 
 static constexpr int launch_bound2 = 4;
 
@@ -51,11 +45,6 @@ __global__ void index_elementwise_kernel(const int64_t N, const func_t f) {
     }
   }
 }
-
-template <int N>
-struct alignas(N) OpaqueType {
-  char data[N];
-};
 
 template <typename Value>
 struct DivMod {
@@ -135,31 +124,6 @@ static OffsetCalculator<N, uint32_t, signed_strides> make_offset_calculator_put(
     std::vector<int64_t> desired_shape, std::array<int64_t*, N> strides_array) {
   return OffsetCalculator<N, uint32_t, signed_strides>(
       desired_shape.size(), desired_shape.data(), strides_array.data());
-}
-
-template <typename IndexT>
-std::array<char*, DDim::kMaxRank> GetIndexDataPtrs(
-    const std::vector<const DenseTensor*> index) {
-  std::array<char*, DDim::kMaxRank> index_ptrs{};
-
-  PADDLE_ENFORCE_LE(index.size(),
-                    DDim::kMaxRank,
-                    "The number of index tensors exceeds the maximum rank.");
-
-  for (size_t i = 0; i < index.size(); ++i) {
-    const IndexT* p_index = index[i]->data<IndexT>();
-
-    PADDLE_ENFORCE_NOT_NULL(
-        p_index,
-        ::common::errors::InvalidArgument(
-            "The pointer p_index is nullptr, "
-            "please check whether the index tensor is valid and "
-            "its data is correctly initialized."));
-
-    index_ptrs[i] = reinterpret_cast<char*>(const_cast<IndexT*>(p_index));
-  }
-
-  return index_ptrs;
 }
 
 template <int N, bool signed_strides = false>
